# PERF: improved clip performance